Various updates to the rwlibs::calibration packages
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4919 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rwlibs/calibration/xml/XmlCalibrationSaver.cpp
@@ -111,7 +111,7 @@ void createDOMDocument(DOMElem::Ptr rootDoc, WorkCellCalibration::Ptr calibratio
 	ElementCreator creator(rootElement);
 
 	CompositeCalibration<FixedFrameCalibration>::Ptr fixedFrameCalibrations = calibration->getFixedFrameCalibrations();
-	for (size_t i = 0; i<fixedFrameCalibrations->getCalibrationCount(); i++) {
+	for (int i = 0; i<fixedFrameCalibrations->getCalibrationCount(); i++) {
 		FixedFrameCalibration::Ptr ffc = fixedFrameCalibrations->getCalibration(i);
 		DOMElem::Ptr element = rootElement->addChild("FixedFrameCalibration");
 		creator.createElement<FixedFrameCalibration::Ptr>(ffc, element);
